pyproject.toml
[project]
name = "hotel-booking-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "jupyter>=1.1.1",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.11",
    "langfuse>=4.0.1",
    "langgraph>=1.1.3",
    "loguru>=0.7.3",
    "openai>=2.29.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pdfplumber>=0.11.0",
    "python-multipart>=0.0.9",
]
